Completed Landing Page
Oops, I totally forgot to commit, I just got into the flow.
I applied font size at the end of the project, so the proportions
are off compared to the desired result. I may choose to change
this in future commits. The current result matches the desired
solution in terms of flex, spacing, and colors when the
page is shrunk.

One learning is that it's important to put font sizes in early.
and another is that CSS specificity also goes against
another system of prioritzaiton for certain styles.